partial_semantic_support has a new argument for vmat
In the last version, vmat was always estimated within the funciton
pyldl.measures.partial_semantic_support every time the function was
called. It does not make sense, of course, from the perspective of
computational loads.

In the current version, the function has a new argument "vmat", which
receives a V-matrix from outside of the function. A V-matrix is
estimated only when the argument "vmat" is not provided, with a message
that tells that the computation may take a while.

On branch master
Changes to be committed:
      modified:   measures.py