Added: "choose_pos" for partial semantic support
In the latest version, pyldl.pyldl.measures.partial_semantic_support did
not accept and give an error when the target trigram is found at
multiple locations within the target word.

In the current version, the function has an additional argument
"choose_pos". The function chooses the one at the specified location by
"choose_pos" when multiple locations are found for the target trigram
within the target word.

On branch master
Changes to be committed:
      modified:   pyldl/measures.py